+++ packages/core/lib/combo/logic/fixer.ts
@@ -2,7 +2,7 @@ import { ItemHelpers, Items } from '../items';
 import { Monitor } from '../monitor';
 import { Settings } from '../settings';
 import { mustStartWithMasterSword } from '../settings/util';
-import { Location, MM_MERCHANTS, MM_SCRUBS, ONE_TIME_SHOP_CHECKS, OOT_FROGS, OOT_MERCHANTS, OOT_ONE_TIME_SCRUBS, isLocationChestFairy, isLocationOtherFairy, makeLocation } from './locations';
+import { Location, MM_MERCHANTS, MM_SCRUBS, ONE_TIME_SHOP_CHECKS, OOT_FROGS, OOT_MERCHANTS, OOT_ONE_TIME_SCRUBS, MM_LOTTERY, isLocationChestFairy, isLocationOtherFairy, makeLocation } from './locations';
 import { World } from './world';
 
 export class LogicPassFixer {
@@ -41,6 +41,10 @@ export class LogicPassFixer {
     }
 
     if(!settings.shuffleFrogsRupeesOot && OOT_FROGS.includes(locId)) {
+      return true;
+    }
+
+    if(!settings.shuffleLotteryMm && MM_LOTTERY.includes(locId)) {
       return true;
     }