+++ packages/core/lib/combo/settings/random.ts
@@ -107,6 +107,7 @@ export async function applyRandomSettings(rnd: OptionRandomSettings, oldSettings
   base.shuffleMerchantsMm = booleanWeighted(random, 0.50);
   base.shuffleMerchantsOot = booleanWeighted(random, 0.50);
   base.shuffleFrogsRupeesOot = booleanWeighted(random, 0.50);
+  base.shuffleLotteryMm = booleanWeighted(random, 0.30);
   base.eggShuffle = booleanWeighted(random, 0.50);
   base.songs = sampleWeighted(random, { songLocations: 6, anywhere: 4 });
   base.divingGameRupeeShuffle = booleanWeighted(random, 0.50);